Refactor ChatActivity (Separation of UI/ViewModel/Data) (#104)
* Refactor files of the ChatActivity and SettingsActivity

- Files placed in the right place
- ChatViewModel still needs to be completed
- NavBar of ChatActivity removed (discussed during meeting)

* Refactor as stated in previous commit

* Implement ChatViewModel and refactor ChatActivity

- Separation of the UI part in ChatActivity and the "logic" in
ChatViewModel
- Note: Some changes are still needed to have perfect separation

* Remove tests for navbar (that was removed in previous commit)

* Fix code according to code review

- ChatMessageListAdaper moved to model package
- getAdapter() split in initAdapter() and getAdapter()
- Other minor changes performed

* Adding comments for use of annotations
