+++ app/src/main/java/com/github/brugapp/brug/model/ChatMessage.kt
@@ -1,6 +1,6 @@
-package com.github.brugapp.brug
+package com.github.brugapp.brug.model
 
-data class ChatItemModel(
+data class ChatMessage(
     val content: String? = null,
     val datetime: String? = null,
     val sender: String? = null
