+++ app/src/main/java/com/github/brugapp/brug/ui/SettingsActivity.kt
@@ -1,7 +1,8 @@
-package com.github.brugapp.brug
+package com.github.brugapp.brug.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.github.brugapp.brug.R
 
 class SettingsActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?){
